requirements.txt
imageio==2.34.0
matplotlib==3.8.2
numpy==1.26.4
opencv_python==4.8.1.78
pandas==2.2.1
Pillow==10.2.0
PyYAML==6.0.1
scipy==1.11.1
einops==0.6.1
# tensorflow==2.15.0.post1
# torch==2.2.1
# torchsummary==1.5.1
# torchview==0.2.6
# torchvision==0.17.1
# tqdm==4.65.0
